Charli.cpp
The C++ version of Addison(A utility tool). They basically just automation for a variety of boring tasks. This one is very rough and clearly as people will see from the unused library's: it has a long way to go, but I've become somewhat infatuated with C++ as a language and felt compelled to write a version of Addison in it. so I present: Charli.