Charli.cpp
Removed redundant use of argv and such within the directory mapper function: and includes that where there because of features that will not be implemented here but in a separate script. Made it a bit more multi-platform so now you(assuming you have the right command line tools installed) can use it on Windows and Linux. Compile with "g++  Charli.cpp -o Charli".